requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pypdf==6.1.1
google-generativeai==0.8.5
numpy==2.3.2
pydantic-settings==2.10.1
streamlit==1.49.1
requests==2.32.5
chromadb==1.0.20
PyMuPDF==1.26.4
